Set html_use_smartypants to False
The html_use_smartypants option converts double dashes into unicode
large dashes. This messes up the table with command line arguments
since a lot of CLAs start with '--'. By setting this to False we can
properly parse the table and copy the argument from the documentation.

If the documentation has rendering issues in other places because of
this option we will have to reconsider using this.