Dev box: Install ROOT 6.08
Upgrade from ROOT 6.06 to 6.08 in the dev box.  This version is
compatible with the C++11 ABI, so the "-D_GLIBCXX_USE_CXX11_ABI=0"
option does not have to be appended to the CXX flags any longer.

A number of patches are applied before building ROOT:

The "SearchInstalledSoftware.cmake" CMake module for ROOT is modified to
avoid linking to libraries in the Anaconda installation.  Linking to
libraries from Anaconda forces users to append the Anaconda library
directory to the LD_LIBRARY_PATH environment variable, which makes
system programs such as "wget", "curl", "ssh", and "gpg" load Anaconda
libraries as well.  This gives problems on Ubuntu, where the system
libraries have the same versions as the Anaconda libraries, but are
slightly different.

JupyROOT (ROOT support for IPython notebooks) currently only works with
Python 2.  Commits which upgrade to Python 3 are cherry-picked from the
ROOT master.  These commits replace relative imports by absolute imports
and remove the use of the __builtin__ module.

The API of TMVA has changed between releases 6.06 and 6.08 and the
ability to set the path for weight files was removed.  This feature is
reimplemented with a TMVA patch.