Fix the --cmd option when running from terminal
The -c / --cmd option did not parse properly in the current project
structure. Now it will be parsed in the run_eskapade script when
checking the --cmd option.

The command can not overwrite existing values in the macro, only add
new values that are going to be used and are not set in the macro.