The class ColumnStats in vis_utils.py has now been renamed and split into two files.
The part of ColumnStats that evaluates statistics of an array is now called
the class ArrayStats, which is stored in analysis/statistics.py
This class has also been updated to accept weights given to records. This makes
is possible to calculate statistics on an histogram.

The plotting functionality of ColumnStats vis_utils.py has been kept in vis_utils,
but is now a stand-alone function.