Reimplement config variables on command line
Configuration variables for the Eskapade run were set on the command
line with the "--cmd" argument of the run_eskapade script.  Options were
given as a string of Python code, in which a user could basically do
anything.

The configuration argument is reimplemented to limit its use to setting
configuration variables.  The "--conf-var" argument can be specified
multiple times as "key=value".  The key-value pairs are set in the
config-object instance for the Eskapade run.

The specified values are interpreted as Python object with the Abstract
Syntax Trees package.  If the interpretation failes, the specified value
is stored as a string.