Roofit code to calculate correlation and residuals of observables
Added roofit pdf RooABCDHistPdf that can be used to test the factorization
assumption of observables in arbitrary number-counting datasets, and to
find the most dominant outliers in this assumptions.

I have committed the following files (that all connect to RooABCDHistPdf):

Classes:

- RooABCDHistPdf.cxx/.h:
RooFit pdf that can be used to test the factorization
assumption of observables in arbitrary number-counting datasets, and to
find the most dominant outliers in this assumptions.

- RooParamHistPdf.cxx/.h
RooABCDHistPdf inherits from RooParamHistPdf, which is a roofit Pdf class
derived from RooHistPdf which multiplies each bin with a RooRealVar scale
parameter, e.g. which can be fit or constrained.

- RooExpandedFitResult.cxx/.h
A class derived from RooFitResult, to be able to add more parameters
for error propagation (for calculation & visualization purposes), or to
build a RooFitResult object from set of parameters

- TMsgLogger.cxx/.h
An output logger class producing nicely formatted log messages, adapted from
TMVA::MsgLogger

Modules:

- ABCDUtils.cxx/.h
A Collection of ROOT utility functions for RooABCDHistPdf. In particular:
A function that returns the (significance of the) p-value of the
hypothesis that the observables in the input dataset
are *not* correlated.
A function returns a dataset with the normalized residuals
(= pull values) for all bins of the input data.

- Statistics.cxx/.h
A collection of ROOT statistics utility functions, based on RooStats and RooFit, namely:
number counting utilities, giving the corresponding p and Z values, and
error propagation function.

- DataUtils.cxx/.h
A collection of ROOT utility functions for processing RooFit RooDataSets:
to convert a roodatahist into a roodataset, and
to calculate the error on a given function for each row in a
RooDataSet, and add this error as a new column to the dataset.