Added new link root_analysis uncorrelation_hypothesis_tester.py
Link calculates and tests for correlations between categorical observables.
Test of correlation between categorical observables taking into account the effects
of low statistics. The test is performed on two levels.

Significance test: tests correlation between observables.
This test calculates the (significance of the) p-value of the hypothesis that
the observables in the input dataset are not correlated. A detailed discription
of the method can be found in ABCDutils.h

Residuals test: tests correlation between values of observables.
This test calculates the normalized residuals (pull values) for each bin in the
dataset, under the hypothesis of no correlation. A detailed discription of the method
can be found in ABCDutils.h

Both tests compare the measured frequency per bin with the expected frequency per bin. The
expected frequency is calculated on the assumption of no correlation. A detailed description
of the calculation of the expected frequency can be found in RooABCDHistPDF.cxx.

Two reports are generated containing the results of the above tests. The long report
(report.tex) contains all the results. The short report (report_client.tex) contains only
the most non-correlating measurements. The results are also saved in the DataStore.

In addition:

In vis_utils.py plot_matrix function, can print both correlation size
and number of events in plot.

Added example macro esk410_testing_correlations_between_categories.py
Macro illustrates use of uncorrelation_hypothesis_tester.py link.
This commit includes integration test of esk410 macro.