ROOT: Do not load RooFit library in links
In two links, the RooFit library was loaded.  Since all links are
imported when the root_analysis package is imported, this load was also
triggered for the Python process of a Spark worker.  This was causing a
crash (why, though?).

To avoid this situation, load the library in the Eskapade macros and not
in links.  This is a temporary work-around and this should be further
investigated and fixed.