ROOT: Load Eskapade RooFit lib at the right place
Load the Eskapade RooFit library in the initialization of a RooFitModel
of a custom RooFit PDF, in the initialize of a link that uses custom
RooFit objects, or in a a macro that uses custom RooFit objects.  Do not
trigger loading the library when the root_analysis package or a module
is imported, as this goes wrong if the library cannot be built/loaded at
that time (for example in a Spark executor).

Additionally, check if all custom RooFit ojbects can be accessed after
building and loading the library.