A simple logging framework for Eskapade.
A logging framework that uses PEP-3101 style formatting and keyword placeholders.
It is essentially a thin layer on top of the Python logging framework. A logger
can be instantiated at the module level or at the class level as a class attribute.

This replaces LoggerMixing and provides an interface that ensures/enforces
consistent logging within Eskapade.

See the code and docs for examples on how to use the logging framework.

Additional changes are:

 * Tox, pytest, flake8 configuration files for CI.

 * Refactoring and cleaning-up of code to comply with PEP-8.

 * String formatting updated to PEP-3101.

 * Fix docstrings to comply with PEP-257.

 * Auto-compilation and auto-loading of esroofit library.

 * Added an eskapade_bootstrap entry point. This can be used to bootstrap
   an Eskapade analysis project.

 * Tests are moved outside of the source folder.

 * Various minor bug fixes.