Load generated roofit library if it exist else exit.
The problem was that the macro assumed that the auto-generated library
and pdf was part of the esroofit library. The macro even encouraged
users to modify the provided esroofit library. This is a no-no. I
removed the message that encouraged users to do this. I also modified
the macro to load the library when it exists else to create. The purpose
of the macro is to illustrate RooFit's on the fly pdf creation and how
it can be used within Eskapade.