Core: introduced new status code BreakChain
Added new status code BreakChain, which is similar to SkipChain but slightly different.
It allows one to break the execution of a chain (and subsequent links), but will still
perform the finalize() of the chain (and all links).

This is useful when looping of an input dataset in chunks and eg filling an histogram,
where the histogram is stored in the finalize() call of a link.