Improved handling of user settings for writers
The handling of the numpy and feather writers have been improved.
For example, if DF was saved with store_index and loaded with don't
restore index, the saved index is stored in the DF with name
'restored_index'.

Readers and writers have been tested as stand alone functions.
Integration into eskapde has to be tested